+++ crates/service/src/algorithms/ivf/ivf_naive.rs
@@ -79,7 +79,7 @@ pub struct IvfRam<S: G> {
     raw: Arc<Raw<S>>,
     quantization: Quantization<S>,
     // ----------------------
-    dims: u16,
+    dims: u32,
     // ----------------------
     nlist: u32,
     // ----------------------
@@ -95,7 +95,7 @@ pub struct IvfMmap<S: G> {
     raw: Arc<Raw<S>>,
     quantization: Quantization<S>,
     // ----------------------
-    dims: u16,
+    dims: u32,
     // ----------------------
     nlist: u32,
     // ----------------------
